requirements.txt
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
jinja2>=3.1.2
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
bcrypt==4.0.1
google-genai>=0.1.0
redis>=5.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
